backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
timeago
pymongo
httpx
prometheus-client
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-logging
opentelemetry-exporter-otlp
opentelemetry-api
opentelemetry-sdk
